move fullscreen nav to spatial controls
