+++ src/lib/components/scatterplot/SpatialControls.js
@@ -4,14 +4,17 @@ import {
   faCrosshairs,
   faDrawPolygon,
   faHand,
+  faList,
   faMinus,
   faPen,
   faPlus,
+  faSearch,
   faSliders,
   faTrash,
 } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { JoinInner } from "@mui/icons-material";
+import useMediaQuery from "@mui/material/useMediaQuery";
 import {
   DrawLineStringMode,
   DrawPolygonByDraggingMode,
@@ -20,6 +23,7 @@ import {
   ModifyMode,
   ViewMode,
 } from "@nebula.gl/edit-modes";
+import { OverlayTrigger, Tooltip } from "react-bootstrap";
 import Button from "react-bootstrap/Button";
 import ButtonGroup from "react-bootstrap/ButtonGroup";
 import Dropdown from "react-bootstrap/Dropdown";
@@ -37,6 +41,8 @@ export function SpatialControls({
   resetBounds,
   increaseZoom,
   decreaseZoom,
+  setShowObs,
+  setShowVars,
 }) {
   const dataset = useDataset();
   const dispatch = useDatasetDispatch();
@@ -44,6 +50,8 @@ export function SpatialControls({
 
   const handleCloseControls = () => setShowControls(false);
   const handleShowControls = () => setShowControls(true);
+  const showObsBtn = useMediaQuery("(max-width: 991.98px)");
+  const showVarsBtn = useMediaQuery("(max-width: 1199.98px)");
 
   const onSelect = (eventKey, event) => {
     switch (eventKey) {
@@ -109,15 +117,26 @@ export function SpatialControls({
   return (
     <div className="cherita-spatial-controls">
       <ButtonGroup vertical className="w-100 mb-1">
-        <Button onClick={increaseZoom} title="Increase zoom">
-          <FontAwesomeIcon icon={faPlus} />
-        </Button>
-        <Button onClick={decreaseZoom} title="Decrease zoom">
-          <FontAwesomeIcon icon={faMinus} />
-        </Button>
-        <Button onClick={resetBounds} title="Reset zoom and center">
-          <FontAwesomeIcon icon={faCrosshairs} />
-        </Button>
+        {showObsBtn && (
+          <OverlayTrigger
+            placement="right"
+            overlay={<Tooltip id="tooltip-obs">Browse categories</Tooltip>}
+          >
+            <Button onClick={() => setShowObs(true)}>
+              <FontAwesomeIcon icon={faList} />
+            </Button>
+          </OverlayTrigger>
+        )}
+        {showVarsBtn && (
+          <OverlayTrigger
+            placement="right"
+            overlay={<Tooltip id="tooltip-vars">Search features</Tooltip>}
+          >
+            <Button onClick={() => setShowVars(true)}>
+              <FontAwesomeIcon icon={faSearch} />
+            </Button>
+          </OverlayTrigger>
+        )}
         <Button onClick={handleShowControls}>
           <FontAwesomeIcon icon={faSliders} />
         </Button>
@@ -129,6 +148,16 @@ export function SpatialControls({
           active={mode === ViewMode}
         >
           <FontAwesomeIcon icon={faHand} />
+        </Button>
+        <Button onClick={increaseZoom} title="Increase zoom">
+          <FontAwesomeIcon icon={faPlus} />
+        </Button>
+        <Button onClick={decreaseZoom} title="Decrease zoom">
+          <FontAwesomeIcon icon={faMinus} />
+        </Button>
+        <div className="border-bottom"></div> {/* Divider */}
+        <Button onClick={resetBounds} title="Reset zoom and center">
+          <FontAwesomeIcon icon={faCrosshairs} />
         </Button>
         <Dropdown
           as={ButtonGroup}
