+++ src/scss/components/plots.scss
@@ -4,4 +4,8 @@
   top: 1rem;
   left: 1rem;
   width: 3rem;
+  .btn {
+    padding-left: 0;
+    padding-right: 0;
+  }
 }
